Add files via upload
The task is to write a program that reads information from a database file containing employees' names, heights, and weights of a company. The program should then print the employees in ascending order of their heights, and in case of equal heights, it should prioritize the person with a lower weight. It is guaranteed that no two individuals have the same height and weight.

The format of the information in the database table is as follows:

Sample Input: (The following information is defined in the database, and this is just an example; the number of individuals may vary)

Height  Weight  Name
180     75      Amin
190     90      Mahdi
175     75      Mohammad
175     60      Ahmad

Output:
Mahdi 190 90
Amin 180 75
Ahmad 175 60
Mohammad 175 75